fix: resolve 11 QA findings — SQL column mismatches, type casts, 404 handling, and missing mock-data module
CRITICAL fixes (6):
- users/route.ts: Fix us.emoji→us.status_emoji, us.text→us.status_text column names
- users/[id]/route.ts: Same column name fix for user profile endpoint
- users/[id]/status/route.ts: Fix UPSERT columns emoji→status_emoji, text→status_text, expiry→expires_at
- channels/[id]/members/route.ts: Add LEFT JOIN user_statuses for display_name (was on users table)
- channels/[id]/pins/route.ts: Fix p.pinned_at→p.created_at in ORDER BY
- channels/[id]/files/route.ts: Fix f.user_id→f.uploaded_by in JOIN

MINOR fixes (4):
- channels/route.ts: Cast COUNT(*)::int for member_count
- channels/[id]/messages/route.ts: Cast COUNT(*)::int for thread_reply_count
- channels/[id]/members/route.ts: Add RETURNING + rowCount check for 404 on DELETE
- workspace/route.ts: Cast COUNT(*)::int for member_count

INFO fix (1):
- src/lib/mock-data.ts: Create typed mock data module matching init-db.sql seed data

+++ src/app/api/channels/[id]/messages/route.ts
@@ -9,7 +9,7 @@ export async function GET(
   try {
     const result = await query(
       `SELECT m.*, u.username, u.avatar_color,
-        (SELECT COUNT(*) FROM threads t WHERE t.parent_message_id = m.id) as thread_reply_count,
+        (SELECT COUNT(*)::int FROM threads t WHERE t.parent_message_id = m.id) as thread_reply_count,
         EXISTS(SELECT 1 FROM pins p WHERE p.message_id = m.id) as is_pinned,
         COALESCE(
           (SELECT json_agg(json_build_object('emoji', r.emoji, 'count', r.cnt, 'user_ids', r.uids))
